# book-network/src/main/java/com/karim/book/feedback/FeedbackController.java
package com.karim.book.feedback;

import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import io.swagger.v3.oas.annotations.tags.Tag;
import lombok.RequiredArgsConstructor;

import org.springframework.security.core.Authentication;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;

import jakarta.validation.Valid;
import jakarta.websocket.server.PathParam;

import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestParam;

import com.karim.book.common.PageResponse;



@RestController
@RequestMapping("feedbacks")
@RequiredArgsConstructor
@Tag(name = "Feedback")
public class FeedbackController {
    private final FeedbackService feedbackService;

    @PostMapping
    public ResponseEntity<Integer> saveFeedback(
        @Valid @RequestBody FeedBackRequest request,
        Authentication connectedUser
    ) {
                
        return ResponseEntity.ok(feedbackService.save(request, connectedUser));
    }
    
    @GetMapping("/book/{book-id}")
    public ResponseEntity<PageResponse<FeedbackResponse>> findAllFeedbacksByBook(
        @PathParam("book-id") Integer bookId,
        @RequestParam(name = "page", defaultValue = "0", required = false ) int page,
        @RequestParam(name = "size", defaultValue = "10", required = false ) int size,
        Authentication connectedUser
    ) {
        return ResponseEntity.ok(feedbackService.findAllFeedbacksByBook(bookId, page, size, connectedUser));
    }
    
}